testsuite: generalized field-merge tests for <32-bit int [PR118025]
Explicitly convert constants to the desired types, so as to not elicit
warnings about implicit truncations, nor execution errors, on targets
whose ints are narrower than 32 bits.


for  gcc/testsuite/ChangeLog

	PR testsuite/118025
	* gcc.dg/field-merge-1.c: Convert constants to desired types.
	* gcc.dg/field-merge-3.c: Likewise.
	* gcc.dg/field-merge-4.c: Likewise.
	* gcc.dg/field-merge-5.c: Likewise.
	* gcc.dg/field-merge-11.c: Likewise.
	* gcc.dg/field-merge-17.c: Don't mess with padding bits.

+++ gcc/testsuite/gcc.dg/field-merge-11.c
@@ -10,7 +10,11 @@ struct s {
   int c;
 } __attribute__ ((aligned (4)));
 
-struct s p = { 42, (short)(0xef1 - 0x1000), 0x12345678 };
+struct s p = {
+  (short)(unsigned short)42,
+  (short)(unsigned short)(0xef1 - 0x1000),
+  (int)(unsigned int)0x12345678
+};
 
 void f (void) {
   if (0
@@ -19,9 +23,9 @@ void f (void) {
       || (int)(signed char)p.b != (int)(signed char)(0xef1 - 0x1000)
       || (unsigned)(unsigned char)p.b != (unsigned)(unsigned char)(0xef1 - 0x1000)
       || (unsigned)p.b != (unsigned short)(0xef1 - 0x1000)
-      || (int)(short)p.b != (int)(0xef1 - 0x1000)
+      || (int)(short)p.b != (int)(short)(unsigned short)(0xef1 - 0x1000)
       || (long)(unsigned char)(p.c >> 8) != (long)(unsigned char)0x123456
-      || p.c != 0x12345678
+      || p.c != (int)(unsigned int)0x12345678
       )
     __builtin_abort ();
 }
